Refactor output_checker.py into galaxy-tool-util.
Needed for remote job completion.
