+++ packages/tool_util/tests/test_output_checker.py
@@ -1,0 +1,1 @@
+../../../test/unit/tools/test_output_checker.py
